kkjnb

+++ Vec3.cpp
@@ -12,10 +12,11 @@
 
 #include "Vec3.hpp"
 
-Vec3 parseVec3(const libconfig::Setting &setting)
-{
-    int x = Vec3::getInt(setting, "x");
-    int y = Vec3::getInt(setting, "y");
-    int z = Vec3::getInt(setting, "z");
+#include "Vec3.hpp"
+
+Vec3 parseVec3(const libconfig::Setting &setting) {
+    float x = Vec3::getFloat(setting, "x");
+    float y = Vec3::getFloat(setting, "y");
+    float z = Vec3::getFloat(setting, "z");
     return Vec3(x, y, z);
 }
